+++ src/array.mjs
@@ -38,3 +38,4 @@ export { default as array_unique } from './array/array_unique.mjs';
 export { default as array_unshift } from './array/array_unshift.mjs';
 export { default as array_walk } from './array/array_walk.mjs';
 export { default as array_walk_recursive } from './array/array_walk_recursive.mjs';
+export { default as in_array } from './array/in_array.mjs';