+++ CHANGELOG.md
@@ -95,6 +95,8 @@
         Рекурсивно применяет пользовательскую функцию к каждому элементу массива
     -   [count](./docs/funcref/vartype/array/func/count.md) &mdash; Подсчитывает количество
         элементов в массиве или в объекте `Countable`
+    -   [in_array](./docs/funcref/vartype/array/func/in_array.md) &mdash; Проверяет, присутствует ли
+        в массиве значение
     -   [array_index_max](./docs/funcref/vartype/array/other/array_index_max.md) &mdash; Возвращает
         наибольший числовой индекс массива.
     -   [array_index_min](./docs/funcref/vartype/array/other/array_index_min.md) &mdash; Возвращает